+++ src/SonsOfPHP/Component/Assert/InvalidArgumentException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SonsOfPHP\Component\Assert;
+
+/**
+ * @author Joshua Estes <joshua@sonsofphp.com>
+ */
+class InvalidArgumentException extends \InvalidArgumentException {}